requirements.txt
aniso8601==9.0.1
aws-wsgi==0.2.7
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==2.0.12
click==8.1.7
firebase-admin==5.2.0
Flask==3.0.1
Flask-RESTful==0.3.9
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
msgpack==1.0.7
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==0.19.2
pytz==2023.3.post1
requests==2.27.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.1
